requirements.txt
-e .

#for i in $(cat requirements.txt); do pip install $i; done
# For testing
pytest==7.2.1
pytest-dotenv==0.5.2

# third-party (libraries)
rake_nltk==1.0.6
ipython==8.10.0

# linting stubs
types-requests==2.28.11.8
types-setuptools==67.1.0.0

# linting
black==22.12.0
isort==5.11.4
mypy==0.991
flake8==6.0.0
flake8-docstrings==1.6.0
pylint==2.15.10

git+https://github.com/openai/whisper.git
pyparsing
newspaper3k
requests
selenium
beautifulsoup4
ffmpeg-python
PyPDF2
gradio
IPython
pydub
ffmpeg
torch
wget
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
tqdm
nltk
gensim
spacy
transformers
pytorch-lightning
pytorch-lightning-bolts
pytorch-lightning-metrics
pytorch-lightning-loggers
pytorch-lightning-tuner
pytorch-lightning-flash
pytorch-lightning-utilities
pytorch-lightning-templates
pytorch-lightning-optimizers
pytorch-lightning-profiler
pytorch-lightning-logging
pytorch-lightning-logging-wandb
pytorch-lightning-logging-tensorboard
pytorch-lightning-logging-neptune
pytorch-lightning-logging-comet
pytube
flask
gpt_index
llama-index
langchain
langchain-utilities
langchain-templates
